+++ app/src/pages/Trends/TrendsExplorePage.js
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types'
 import * as qs from 'query-string'
 import { Button } from 'semantic-ui-react'
 import Selector from './../../components/Selector/Selector'
+import GetTimeSeries from './../../components/GetTimeSeries'
 import TrendsExploreHeader from '../../components/Trends/Explore/TrendsExploreHeader'
 import GetTrends from './../../components/Trends/GetTrends'
-import GetTimeSeries from './../../components/GetTimeSeries'
 import TrendsReChart from './../../components/Trends/TrendsReChart'
+import TrendsStats from './../../components/Trends/TrendsStats'
 import SmoothDropdown from '../../components/SmoothDropdown/SmoothDropdown'
 import SmoothDropdownItem from '../../components/SmoothDropdown/SmoothDropdownItem'
 import ShareOptions from './../../components/ShareOptions/ShareOptions'
-import GetTrendsStats from './../../components/Trends/GetTrendsStats'
 import { capitalizeStr } from './../../utils/utils'
 import './TrendsExplorePage.css'
 
@@ -24,35 +24,6 @@ export const getStateFromQS = ({ location }) => {
     asset: asset || 'bitcoin'
   }
 }
-
-export const Stats = () => (
-  <div>
-    <GetTrendsStats
-      render={({ stats, isLoading }) => {
-        if (isLoading) {
-          return ''
-        } else {
-          return (
-            <div style={{ background: '#f5f9ff99' }}>
-              <hr />
-              <strong style={{ marginLeft: 40 }}>Our statistic:</strong>
-              <div style={{ display: 'flex', marginTop: 0 }}>
-                <ul>
-                  <li>All messages: {stats.documentsCount}</li>
-                  <li>Messages per day: {stats.averageDocumentsPerDay}</li>
-                </ul>
-                <ul>
-                  <li>Telegram channels: {stats.telegramChannelsCount}</li>
-                  <li>Subreddits: {stats.subredditsCount}</li>
-                </ul>
-              </div>
-            </div>
-          )
-        }
-      }}
-    />
-  </div>
-)
 
 export class TrendsExplorePage extends Component {
   state = {
@@ -133,7 +104,7 @@ export class TrendsExplorePage extends Component {
               />
             )}
           />
-          <Stats />
+          <TrendsStats timeRange={timeRange} />
         </div>
       </div>
     )